refactor: StopPropagation
